Applied Grid System, image still needs fitting
I dont like css modules at the current moment, cards flip and content is on backside. Now I just need to get the image to fit in a way I like and add more content to the back. Maybe add a button that will take you to the instructions if the api provides that